backend/requirements-dev.txt
pytest==8.2.2
pytest-mock==3.14.0
moto[dynamodb]==5.0.12
boto3-stubs[dynamodb,cognito-idp]==1.34.144
